Migrate to Neovim 0.5
The newer version of Neovim (0.5) [[1]] has an improved performance and
has a native LSP support. LuaJIT is also integrated to Neovim which
provides a better and faster way to do things.

There are a few plugins worth mentioning:
1. Neovim [LSP]
  A collection of common configuration for Neovim' builtin LSP.
2. [Treesitter]
  A syntax highligher. it's really faster and makes code beautiful!
3. [Telescope]
  A fuzzy finder built on latest neovim features.

Also, some plugins are replaced by their Neovim 0.5 alternatives which
is developed in Lua.

[1]: https://github.com/neovim/neovim
[Treesitter]: https://github.com/nvim-treesitter/nvim-treesitter
[Telescope]: https://github.com/nvim-telescope/telescope.nvim
[LSP]: https://github.com/neovim/nvim-lspconfig